finalizacao dos controller e insercao das configuracoes do docker

+++ src/main/java/com/gustavoaviila/Delivery/domain/entity/Produto.java
@@ -1,20 +1,14 @@
 package com.gustavoaviila.Delivery.domain.entity;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
-import java.util.Set;
 
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -37,5 +31,5 @@ public class Produto implements Serializable {
     private String descricao;
 
     @NotNull(message = "{campo.preco.obrigatorio}")
-    private BigDecimal preco;
+    private Double preco;
 }
